LayoutTests/media/audio-constructor-autobuffer-expected.txt

Test that Audio() sets 'autobuffer' attribute.

RUN(mediaElement = new Audio())

++ Test initial attribute value
EXPECTED (mediaElement.getAttribute('autobuffer') != 'null') OK
EXPECTED (mediaElement.autobuffer == 'true') OK

RUN(mediaElement.src = 'content/test.wav')
RUN(mediaElement.load())

++ Test after 'canplaythrough' event
EXPECTED (mediaElement.getAttribute('autobuffer') != 'null') OK
EXPECTED (mediaElement.autobuffer == 'true') OK
END OF TEST
